Init aws client from automationtoken (#47)
* Init aws client from automationtoken

AWS API clients need to be initialized with Vault Credentials in order to be compatible with existing infrastructure.

* Make it possible to pass in an AWS client to NewS3State, this is required for testing outside of the package

* Fixes, make query package local, fix logging

* Update template

* Refactoring, remove dependencies from AWS Client

Refactoring code to pass in credentials into AWS Client instead of
having them resolved by the Client itself. This makes the code easier to
read and extend, i.e. if we'll want to use other ways of passing in
credienials.

* Remove comments and newline

* PR review remarks